Split account-related intercepts into its own file
