+++ packages/manager/cypress/support/intercepts/profile.ts
@@ -1,5 +1,5 @@
 /**
- * @file Cypress intercepts and mocks for Cloud Manager profile and account requests.
+ * @file Cypress intercepts and mocks for Cloud Manager profile requests.
  */
 
 import { makeErrorResponse } from 'support/util/errors';
@@ -8,17 +8,6 @@ import type {
   SecurityQuestionsData,
   SecurityQuestionsPayload,
 } from '@linode/api-v4/types';
-
-/**
- * Intercepts GET request to fetch account user information.
- *
- * @param username - Username of user whose info is being fetched.
- *
- * @returns Cypress chainable.
- */
-export const interceptGetUser = (username: string): Cypress.Chainable<null> => {
-  return cy.intercept('GET', `*/account/users/${username}`);
-};
 
 /**
  * Intercepts GET request to fetch user profile.
@@ -148,29 +137,5 @@ export const mockConfirmTwoFactorAuth = (
 ): Cypress.Chainable<null> => {
   return cy.intercept('POST', '*/profile/tfa-enable-confirm', {
     scratch: scratchCode,
-  });
-};
-
-/**
- * Intercepts PUT request to update account username and mocks response.
- *
- * @param oldUsername - The original username which will be changed.
- * @param newUsername - The new username for the account.
- * @param restricted - Whether or not the account is restricted.
- *
- * @returns Cypress chainable.
- */
-export const mockUpdateUsername = (
-  oldUsername: string,
-  newUsername: string,
-  restricted: boolean = false
-) => {
-  return cy.intercept('PUT', `*/account/users/${oldUsername}`, {
-    username: newUsername,
-    email: 'mockEmail@example.com',
-    restricted,
-    ssh_keys: [],
-    tfa_enabled: false,
-    verified_phone_number: null,
   });
 };
